api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
supabase==2.3.4
postgrest==0.13.0
realtime==0.0.2
httpx==0.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
